+++ macos/Runner/MainFlutterWindow.swift
@@ -1,5 +1,6 @@
 import Cocoa
 import FlutterMacOS
+import desktop_multi_window
 
 class MainFlutterWindow: NSWindow {
   override func awakeFromNib() {
@@ -9,6 +10,11 @@ class MainFlutterWindow: NSWindow {
     self.setFrame(windowFrame, display: true)
 
     RegisterGeneratedPlugins(registry: flutterViewController)
+
+    FlutterMultiWindowPlugin.setOnWindowCreatedCallback { controller in
+      // Register the generated plugins for each new window's engine.
+      RegisterGeneratedPlugins(registry: controller)
+    }
 
     super.awakeFromNib()
   }